Use BaseJdbcConnectorTest for Phoenix connectors
